Fixed overflow of 64-bit integers read from database

+++ src/wp/SQLite.cs
@@ -1808,7 +1808,7 @@ namespace SQLite
                                 targetType = typeof(string);
                                 break;
                             case SQLite3.ColType.Integer:
-                                targetType = typeof(int);
+                                targetType = typeof(long);
                                 break;
                             case SQLite3.ColType.Float:
                                 targetType = typeof(double);
